# packages/contract/openapi.json
{
  "components": {
    "schemas": {
...
      "Body_perfis_upload_banner": {
        "properties": {
          "file": {
            "contentMediaType": "application/octet-stream",
            "description": "PNG, JPG ou WebP até 5 MB",
            "title": "File",
            "type": "string"
          },
          "version": {
            "minimum": 1.0,
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "file",
          "version"
        ],
        "title": "Body_perfis_upload_banner",
        "type": "object"
      },
      "Body_perfis_upload_logo": {
        "properties": {
          "file": {
            "contentMediaType": "application/octet-stream",
            "description": "PNG, JPG ou WebP até 5 MB",
            "title": "File",
            "type": "string"
          },
          "version": {
            "minimum": 1.0,
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "file",
          "version"
        ],
        "title": "Body_perfis_upload_logo",
        "type": "object"
      },
...
      "CreatePerfilIn": {
        "properties": {
          "bio": {
            "default": "",
            "maxLength": 2000,
            "title": "Bio",
            "type": "string"
          },
          "language": {
            "default": "pt-BR",
            "pattern": "^[a-z]{2,3}(-[A-Z]{2})?$",
            "title": "Language",
            "type": "string"
          },
          "name": {
            "maxLength": 80,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "niche": {
            "default": "",
            "maxLength": 200,
            "title": "Niche",
            "type": "string"
          },
          "slug": {
            "maxLength": 60,
            "minLength": 2,
            "pattern": "^[a-z0-9]+(-[a-z0-9]+)*$",
            "title": "Slug",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/PerfilStatus",
            "default": "em_preparacao"
          }
        },
        "required": [
          "name",
          "slug"
        ],
        "title": "CreatePerfilIn",
        "type": "object"
      },
...
      "ErrorBody": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorBody",
        "type": "object"
      },
      "ErrorEnvelope": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorEnvelope",
        "type": "object"
      },
...
      "ImageRef": {
        "properties": {
          "height": {
            "title": "Height",
            "type": "integer"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "urls": {
            "$ref": "#/components/schemas/ImageUrls"
          },
          "width": {
            "title": "Width",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "width",
          "height",
          "urls"
        ],
        "title": "ImageRef",
        "type": "object"
      },
      "ImageUrls": {
        "properties": {
          "medium": {
            "title": "Medium",
            "type": "string"
          },
          "thumb": {
            "title": "Thumb",
            "type": "string"
          }
        },
        "required": [
          "thumb",
          "medium"
        ],
        "title": "ImageUrls",
        "type": "object"
      },
...
      "Perfil": {
        "properties": {
          "archived": {
            "title": "Archived",
            "type": "boolean"
          },
          "archivedAt": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Archivedat"
          },
          "banner": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ImageRef"
              },
              {
                "type": "null"
              }
            ]
          },
          "bio": {
            "title": "Bio",
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "title": "Createdat",
            "type": "string"
          },
          "createdBy": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/UserRef"
              },
              {
                "type": "null"
              }
            ]
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "language": {
            "title": "Language",
            "type": "string"
          },
          "logo": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ImageRef"
              },
              {
                "type": "null"
              }
            ]
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "niche": {
            "title": "Niche",
            "type": "string"
          },
          "platforms": {
            "items": {
              "$ref": "#/components/schemas/Platform"
            },
            "title": "Platforms",
            "type": "array"
          },
          "slug": {
            "title": "Slug",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/PerfilStatus"
          },
          "updatedAt": {
            "format": "date-time",
            "title": "Updatedat",
            "type": "string"
          },
          "updatedBy": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/UserRef"
              },
              {
                "type": "null"
              }
            ]
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "slug",
          "name",
          "niche",
          "bio",
          "language",
          "status",
          "logo",
          "banner",
          "platforms",
          "archived",
          "archivedAt",
          "version",
          "createdAt",
          "updatedAt",
          "createdBy",
          "updatedBy"
        ],
        "title": "Perfil",
        "type": "object"
      },
...
      "PerfilOut": {
        "properties": {
          "perfil": {
            "$ref": "#/components/schemas/Perfil"
          }
        },
        "required": [
          "perfil"
        ],
        "title": "PerfilOut",
        "type": "object"
      },
      "PerfilStatus": {
        "enum": [
          "em_preparacao",
          "ativo",
          "pausado"
        ],
        "title": "PerfilStatus",
        "type": "string"
      },
      "PerfisList": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/Perfil"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "title": "PerfisList",
        "type": "object"
      },
      "Platform": {
        "enum": [
          "tiktok",
          "youtube",
          "instagram",
          "kwai",
          "facebook",
          "x",
          "outra"
        ],
        "title": "Platform",
        "type": "string"
      },
...
      "UserRef": {
        "properties": {
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name"
        ],
        "title": "UserRef",
        "type": "object"
      },
...
      "VersionIn": {
        "properties": {
          "version": {
            "minimum": 1.0,
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "version"
        ],
        "title": "VersionIn",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "SociMan API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/perfis": {
      "get": {
        "operationId": "perfis_list",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 80,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/PerfilStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "archived",
            "required": false,
            "schema": {
              "default": false,
              "title": "Archived",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PerfisList"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Dados inválidos (validation_error)"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Forbidden"
          }
        },
        "summary": "List Perfis"
      },
      "post": {
        "operationId": "perfis_create",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePerfilIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PerfilOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Forbidden"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Conflict"
          }
        },
        "summary": "Create Perfil"
      }
    },
...
    "/api/perfis/{perfil_id}/banner": {
      "put": {
        "operationId": "perfis_upload_banner",
        "parameters": [
          {
            "in": "path",
            "name": "perfil_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Perfil Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_perfis_upload_banner"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PerfilOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Conflict"
          }
        },
        "summary": "Upload Banner"
      }
    },
    "/api/perfis/{perfil_id}/banner/clear": {
      "post": {
        "operationId": "perfis_clear_banner",
        "parameters": [
          {
            "in": "path",
            "name": "perfil_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Perfil Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VersionIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PerfilOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Conflict"
          }
        },
        "summary": "Clear Banner"
      }
    },
...
    "/api/perfis/{perfil_id}/logo": {
      "put": {
        "operationId": "perfis_upload_logo",
        "parameters": [
          {
            "in": "path",
            "name": "perfil_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Perfil Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_perfis_upload_logo"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PerfilOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Conflict"
          }
        },
        "summary": "Upload Logo"
      }
    },
    "/api/perfis/{perfil_id}/logo/clear": {
      "post": {
        "operationId": "perfis_clear_logo",
        "parameters": [
          {
            "in": "path",
            "name": "perfil_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Perfil Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VersionIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PerfilOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Conflict"
          }
        },
        "summary": "Clear Logo"
      }
    },
...
  }
}